Wire driver threads, command dispatch, and control loop
- Driver threads: load Lua, init, poll loop with backoff and watchdog
- Command dispatch: mpsc channels from control loop to driver threads
- Main: full thread orchestration with graceful shutdown
- Config: Clone derives for thread-safe driver config passing
- Tested live: Ferroamp MQTT driver reading real telemetry,
  API returning live grid/PV/battery data at /api/status

+++ src/lua/host_api/modbus.rs
@@ -104,7 +104,7 @@ fn register_write_multiple(
     let f = lua.create_function(move |_, (register, values): (u16, Table)| {
         // Convert Lua table (1-indexed) to Vec<u16>
         let mut vals = Vec::new();
-        let len: i64 = values.raw_len();
+        let len: i64 = values.raw_len() as i64;
         for i in 1..=len {
             let v: u16 = values.get(i)?;
             vals.push(v);
